Switching to active state when calling startScanning
This is the only exception that does not go over the setState function in ScanditSDK.java. Not being in the active state caused issues in the didScan callback, as mShouldBlockForDidScan was not true, so it actually did not block until the callback ended, resulting in problems like failing to reject codes. (#79)